Modifications to make the exports easier to manage.
Each day for each room is now exported into a separate directory. If you have a large number of uploads (say, from back when LOLcats were extra-popular), this places the files uploads into a single day's directory for easier access.

Both XML and plaintext transcripts are now saved.

Directories are skipped if there were no messages posted that day.

Room names, instead of ID numbers, are used for the top of the export directory tree.

There is both a start and end date, and the latter is now inclusive, for better export control.

The progress messages are a little more informative.

While the 37signals' API limits are generous, I added a 1/10th of a second sleep between each days' requests to ensure we don't go over.