Clean up file export and log export errors; add HTML export.
This mostly just refactors the export loop to be a little tighter and cleaner, while adding some error handling and reporting. Uploaded files are now downloaded using the 'upload' API method, instead of scraping them out of the HTML transcript.

The transcript is exported in three formats: XML, HTML (which is new), and plaintext.

The export path now includes the Campfire site subdomain, so the script could be run non-destructively on several sites from the same directory.