Fix serious name collisions in exported files.
It appears there's a bug in the Campfire API, causing the filenames of uploaded files to be reported incorrectly. Files of the form 'filename-#+.ext' are being reported as 'filename.ext'. If the transcript had the same filename uploaded multiple times, Campfire renames subsequent copies as 'filename-1.ext', 'filename-2.ext', and so on, which would cause all of them to be saved as 'filename.ext'. This would overwrite the original upload with each subsequent upload using that naming pattern, destroying one or more copies in the process.

The transcript XML (and the transcript HTML) both report the original filename correctly; only the `/upload` method is affected. As a result, this uses the transcript XML's version of the filename to save out each file.  With this approach no data is overwritten.

I'm collecting examples of the problem and then will report it to 37signals. Hopefully they can fix it and this can be undone. That said, some checking of the export's accuracy is probably a really good idea, since I imagine some people who use this script will destroy their Campfire accounts after running it.