Break out an upload object and make it not overwrite other files.
Makes an upload object out of some scattered upload code, and resolves problems with uploads overwriting each other. The source of these was the same filename used by more than one upload in a day. The heavy-handed solution is to prepend the upload filename with the upload ID, so it is always unique.

Also, breaks out the config into a home directory YAML file so I can more easily test.

This remove the export statistics and verification, which will need to be added back in later in another form.