Avoid overwriting uploads with the same name by writing them to subdirs.
If a particular file, e.g., `Picture 1.png`, is uploaded to the same room more than once in a day, it would be overwritten were they put in the same directory. While uploads show up in the transcript as `Picture 1.png`, `Picture 1-1.png`, `Picture 1-2.png`, and so on, the filename remains the same in the API.

This could be solved by replicating the versioned naming scheme described above, but those versioned filenames would then no longer match the filenames in the HTML transcript URLs. Instead, I've chosen to write uploads to a subdirectory of the transcript export directory, using the scheme: `uploads/#{id}/#{filename}`.

This scheme matches the tail of the upload URL in the HTML transcript, so a simple text substitution in the HTML transcript can be used to make the uploads live links in those transcripts.