More sophisticated date range checking.
Export will now start on the user's requested start date, or the creation date of the Campfire room, whichever is later; and will end of the user's requested end date, or the date of the last comment in the Campfire room, whichever is earlier. If the user does not provide a start or end date, the full history of the room, from the creation date to the last comment, will be exported.