Make time zones obey my commands.
The Campfire API does not expose the user's set time zone, but transcript requests are localtime for the account, while timestamps are GMT. What a mess!

This adds a download of the user's HTML settings page - which shows the user's selected time zone. Unfortunately, that time zone is stated in Rails time zone format, not Ruby/TZInfo format. More mess!

And, the Rails timezone support is in active_support, which currently has a bug in it and which has all sorts of gem version conflicts and deprecation notices.  Still more mess!

So this avoids all the mess by copying & pasting the Rails <-> TZInfo mapping information out of active_support and including just the part we need here. Not ideal but ....whatever. I don't want to require a buggy active_support gem that could cause people to have to upgrade Rails and/or RubyGems just to use this code.

And, it works.