More detail in the README.
I hate finding something useful and realizing I have to load in a whole set of jargon just to make use of it. Since I'm assuming some people not already friendly with gems might want to export their Campfire accounts, I'm trying to make this as easy to follow as I can.