Clean up exception handling and logging.
I think this puts all of the exception handling at the "right" level -- bubbling up to a meaningful report, while still allowing the overall export to proceed if possible.  Error logs are now written to STDERR, while progress logs go to STDOUT.

At this point I am really missing tests and want to go rig some up if I do much more on this project.

There's a questionable design decision in here, which is that timezone defaults to GMT if I can't get it otherwise; that destroys data in some sense since the plaintext transcript timestamps and the export range requests will be wrong. In my defense it should be the very first line logged when the export runs, so hopefully it will at least be noticeable.

Fixes #5.